refactor(cross-chain): consolidate discovery tests and add currencies request schema
Move discovery tests from discovery.spec.ts into provider.spec.ts and
schemas.spec.ts. Add RelayCurrenciesRequestSchema for POST /currencies/v2
request body validation per official API documentation.
